Chore: 타입분리 및 경로 변경 등 자잘한 수정
