+++ src/types/user.ts
@@ -17,3 +17,15 @@ export interface CommentUser {
   id: string;
   nickname: string;
 }
+
+export interface PostDetail {
+  id: string;
+  createdAt: string;
+  updatedAt: string;
+  title: string;
+  viewCount: number;
+  commentCount: number;
+  isAuthor: boolean;
+  content: string;
+  author: User;
+}
